Move UserProfile creation to Login Screen
