+++ memeolist/app/src/main/AndroidManifest.xml
@@ -13,16 +13,16 @@
         android:roundIcon="@mipmap/ic_launcher_round"
         android:supportsRtl="true"
         android:theme="@style/AppTheme">
-        <activity android:name=".ui.MemeListActivity">
+        <activity
+            android:name=".ui.LoginActivity"
+            android:theme="@style/AppTheme.NoActionBar">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
 
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
         </activity>
-        <activity
-            android:name=".ui.LoginActivity"
-            android:theme="@style/AppTheme.NoActionBar" />
+        <activity android:name=".ui.MemeListActivity" />
         <activity android:name=".ui.MemeFormActivity" />
         <activity android:name=".ui.CommentsFormActivity" />
         <activity
