+++ memeolist/app/src/main/java/org/aerogear/android/app/memeolist/model/UserProfile.java
@@ -1,12 +1,12 @@
 package org.aerogear.android.app.memeolist.model;
 
 import org.aerogear.android.app.memeolist.graphql.AllMemesQuery;
+import org.aerogear.android.app.memeolist.graphql.CreateProfileMutation;
 import org.aerogear.android.app.memeolist.graphql.MemeAddedSubscription;
-import org.aerogear.mobile.auth.user.UserPrincipal;
+import org.aerogear.android.app.memeolist.graphql.ProfileQuery;
 import org.jetbrains.annotations.NotNull;
 
 import java.io.Serializable;
-import java.util.List;
 
 /**
  */
@@ -16,13 +16,6 @@ public class UserProfile implements Serializable {
     private final String displayName;
     private final String email;
     private final String pictureUrl;
-
-    public UserProfile(UserPrincipal loggedUser) {
-        this.id = "1";
-        this.displayName = loggedUser.getName();
-        this.email = loggedUser.getEmail();
-        this.pictureUrl = "https://randomuser.me/api/portraits/lego/1.jpg";
-    }
 
     public UserProfile(String id, String displayName, String email, String pictureUrl) {
         this.id = id;
@@ -55,4 +48,11 @@ public class UserProfile implements Serializable {
         return new UserProfile(user.id(), user.displayname(), user.email(), user.pictureurl());
     }
 
+    public static UserProfile from(ProfileQuery.Profile user) {
+        return new UserProfile(user.id(), user.displayname(), user.email(), user.pictureurl());
+    }
+
+    public static UserProfile from(CreateProfileMutation.CreateProfile user) {
+        return new UserProfile(user.id(), user.displayname(), user.email(), user.pictureurl());
+    }
 }
